sdk/py-sdk/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
# Use the local Python SDK
-e ../../../fern-python-sdk
